mirror/requirements.txt
Django==3.2.7
asgiref==3.4.1
mysqlclient==2.0.3
django-mysql==4.0.0
djangorestframework==3.12.4
pytz==2021.1
setuptools==57.4.0
sqlparse==0.4.1
wheel==0.37.0
